src/main/java/oop2/module01/Person_Solution.java
package oop2.module01;

import java.util.Objects;

/**
 * @author Dieter Holz
 */
public class Person_Solution {
	private int    age;
	private double weight;
	private double height;

    private Person_Solution[] friends;

	public Person_Solution(int age, double weight, double height) {
		this.age    = age;
		this.weight = weight;
		this.height = height;

        friends = new Person_Solution[10];
    }

	public double getBMI() {
		return weight / (height * height);
	}

    public void newFriend(Person_Solution friend) {
	    if(hasAsFriend(friend)){
            return;
        }

        for (int i = 0; i < friends.length; i++) {
            if(friends[i] == null){
                friends[i] = friend;
                break;
            }
        }

        friend.newFriend(this);
    }

    public boolean hasAsFriend(Person_Solution person) {
        for (Person_Solution f : friends) {
            if (Objects.equals(f, person)) {
                return true;
            }
        }
        return false;
    }

    public int numberOfFriends() {
        int counter = 0;
        for (Person_Solution f : friends) {
            if (f != null) {
                counter++;
            }
        }
        return counter;
    }

    public void noFriend(Person_Solution friend) {
	    if(hasAsFriend(friend)){
            for (int i = 0; i < friends.length; i++) {
                if(Objects.equals(friends[i], friend)){
                    friends[i] = null;
                    break;
                }
            }
            friend.noFriend(this);
        }
    }

	//alle Getter und Setter

	public int getAge() {
		return age;
	}

	public void setAge(int age) {
		this.age = age;
	}

	public double getWeight() {
		return weight;
	}

	public void setWeight(double weight) {
		this.weight = weight;
	}

	public double getHeight() {
		return height;
	}

	public void setHeight(double height) {
		this.height = height;
	}


}
